backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.7",
    "python-jose[cryptography]>=3.5.0",
    "python-telegram-bot>=22.4",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "twilio>=9.8.1",
    "uvicorn>=0.35.0",
]
